reduce direct dependencies
