+++ shoebox/app/com/keepit/controllers/ext/ExtUserController.scala
@@ -1,5 +1,6 @@
 package com.keepit.controllers.ext
 
+import com.keepit.controllers.core.SliderInfoLoader
 import com.keepit.classify.{Domain, DomainClassifier, DomainRepo, DomainStates}
 import com.keepit.common.db._
 import com.keepit.common.db.slick._
@@ -26,52 +27,21 @@ import com.google.inject.{Inject, Singleton}
 
 @Singleton
 class ExtUserController @Inject() (db: Database,
-  normalizedURIRepo: NormalizedURIRepo, domainRepo: DomainRepo, userToDomainRepo: UserToDomainRepo, socialConnectionRepo: SocialConnectionRepo, userRepo: UserRepo,
-  basicUserRepo: BasicUserRepo, followRepo: FollowRepo, bookmarkRepo: BookmarkRepo, commentRepo: CommentRepo,
-  domainClassifier: DomainClassifier, uriGraph: URIGraph, userWithSocialRepo: UserWithSocialRepo)
+  domainRepo: DomainRepo, userToDomainRepo: UserToDomainRepo, socialConnectionRepo: SocialConnectionRepo, userRepo: UserRepo,
+  basicUserRepo: BasicUserRepo, sliderInfoLoader: SliderInfoLoader)
     extends BrowserExtensionController {
 
   def getSliderInfo(url: String) = AuthenticatedJsonAction { request =>
-    val userId = request.userId
-    val (bookmark, following, socialUsers, numComments, numMessages, neverOnSite, sensitive) = db.readOnly {implicit s =>
-      val nUri = normalizedURIRepo.getByNormalizedUrl(url)
-      val host: Option[String] = URI.parse(url).get.host.map(_.name)
-      val domain: Option[Domain] = host.flatMap(domainRepo.get(_))
-      val neverOnSite: Option[UserToDomain] = domain.flatMap { domain =>
-        userToDomainRepo.get(userId, domain.id.get, UserToDomainKinds.NEVER_SHOW)
-      }
-      val sensitive: Option[Boolean] =
-        domain.flatMap(_.sensitive).orElse(host.flatMap(domainClassifier.isSensitive(_).right.toOption))
-
-      nUri match {
-        case Some(uri) =>
-          val bookmark = bookmarkRepo.getByUriAndUser(uri.id.get, userId)
-          val following = followRepo.get(userId, uri.id.get).isDefined
-
-          val friendIds = socialConnectionRepo.getFortyTwoUserConnections(userId)
-          val searcher = uriGraph.getURIGraphSearcher
-          val friendEdgeSet = searcher.getUserToUserEdgeSet(userId, friendIds)
-          val sharingUserIds = searcher.intersect(friendEdgeSet, searcher.getUriToUserEdgeSet(uri.id.get)).destIdSet - userId
-          val socialUsers = sharingUserIds.map(u => userWithSocialRepo.toUserWithSocial(userRepo.get(u))).toSeq
-
-          val numComments = commentRepo.getPublicCount(uri.id.get)
-          val numMessages = commentRepo.getMessages(uri.id.get, userId).size
-
-          (bookmark, following, socialUsers, numComments, numMessages, neverOnSite, sensitive)
-        case None =>
-          (None, false, Nil, 0, 0, neverOnSite, sensitive)
-      }
-    }
-
+    val slicerInfo = sliderInfoLoader.load(request.userId, url)
     Ok(JsObject(Seq(
-        "kept" -> JsBoolean(bookmark.isDefined),
-        "private" -> JsBoolean(bookmark.map(_.isPrivate).getOrElse(false)),
-        "following" -> JsBoolean(following),
-        "friends" -> userWithSocialSerializer.writes(socialUsers),
-        "numComments" -> JsNumber(numComments),
-        "numMessages" -> JsNumber(numMessages),
-        "neverOnSite" -> JsBoolean(neverOnSite.isDefined),
-        "sensitive" -> JsBoolean(sensitive.getOrElse(false)))))
+        "kept" -> JsBoolean(slicerInfo.bookmark.isDefined),
+        "private" -> JsBoolean(slicerInfo.bookmark.map(_.isPrivate).getOrElse(false)),
+        "following" -> JsBoolean(slicerInfo.following),
+        "friends" -> userWithSocialSerializer.writes(slicerInfo.socialUsers),
+        "numComments" -> JsNumber(slicerInfo.numComments),
+        "numMessages" -> JsNumber(slicerInfo.numMessages),
+        "neverOnSite" -> JsBoolean(slicerInfo.neverOnSite.isDefined),
+        "sensitive" -> JsBoolean(slicerInfo.sensitive.getOrElse(false)))))
   }
 
   def suppressSliderForSite() = AuthenticatedJsonAction { request =>
